changed marker color by category and changed layout
category buttons on the category bar has different colors, and the makers of each category has corresponding colors.
if a marker of a category is clicked, the corresponding category box is displayed on the left.
deleted blank-cateogry-box and displayed all restaurant elements on default.
changed whole page layout. fixed the left scrollable div from vw to pixels. fixed the height of the elements to disable scrolls of whole page.